66141: SystemML Engine Enhancements - Rework runtime integer scalars for robustness/consistency (long values, error handling overflows), incl tests

+++ SystemML/SystemML/DML/src/main/java/com/ibm/bi/dml/api/jmlc/PreparedScript.java
@@ -102,7 +102,7 @@ public class PreparedScript
 	 * @param scalar
 	 * @throws DMLException 
 	 */
-	public void setScalar(String varname, int scalar) 
+	public void setScalar(String varname, long scalar) 
 		throws DMLException
 	{
 		if( !_inVarnames.contains(varname) )
