+++ SystemML/SystemML/DML/src/main/java/com/ibm/bi/dml/runtime/functionobjects/Modulus.java
@@ -47,21 +47,21 @@ public class Modulus extends ValueFunction
 	}
 
 	@Override
-	public double execute(double in1, int in2) {
+	public double execute(double in1, long in2) {
 		if( in2==0 )
 			return Double.NaN;
 		return in1 - _intdiv.execute(in1, in2)*in2;
 	}
 
 	@Override
-	public double execute(int in1, double in2) {
+	public double execute(long in1, double in2) {
 		if( in2==0.0 || in2==-0.0 )
 			return Double.NaN;
 		return in1 - _intdiv.execute(in1, in2)*in2;
 	}
 
 	@Override
-	public double execute(int in1, int in2) {
+	public double execute(long in1, long in2) {
 		if( in2==0 )
 			return Double.NaN;
 		return in1 - _intdiv.execute(in1, in2)*in2;
